backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
httpx==0.27.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.31.0
aiohttp==3.9.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
alembic==1.13.1